Move example to their own file
